pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mesa"
version = "0.1.0"
description = "Memory Eval Suite for Agents — a reproducible framework for personal AI memory systems"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "rouge-score>=0.1.2",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7", "jsonschema>=4"]
chroma = ["chromadb>=0.4"]
mem0 = ["mem0ai>=0.1"]

[project.scripts]
mesa = "mesa.runner:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mesa*", "adapters*"]
